requirements.txt
# Core dependencies
SpeechRecognition==3.10.0
gTTS==2.3.2
pywhatkit==5.4.3
PyAudio==0.2.11
playsound==1.3.0

# Web and API
requests==2.31.0
urllib3==2.0.7

# Configuration and utilities
PyYAML==6.0.1
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Audio processing
pydub==0.25.1

# System utilities
psutil==5.9.6

# Optional: For better audio quality on macOS
sounddevice==0.4.6
